pyproject.toml
[project]
name = "pronatura-waste-collection-schedule-exporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
